move drawer code for tile-block-drawable to its own file
