+++ src/duckling/game/drawable/drawable-drawer.ts
@@ -28,7 +28,8 @@ import {Drawable, DrawableType, cppTypeToDrawableType} from './drawable';
 import {ShapeDrawable} from './shape-drawable';
 import {ContainerDrawable} from './container-drawable';
 import {ImageDrawable} from './image-drawable';
-import {TileBlockDrawable, getTileHeight, getTileWidth, drawTileBlockDrawable} from './tile-block-drawable';
+import {TileBlockDrawable, getTileHeight, getTileWidth} from './tile-block-drawable';
+import {drawTileBlockDrawable} from './tile-block-drawable-drawer';
 import {AnimatedDrawable} from './animated-drawable';
 import {TextDrawable} from './text-drawable';
 import {ShapeType, Shape, cppTypeToShapeType} from './shape';
